+++ core/core.pl
@@ -3384,7 +3384,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 % versions, 'rcN' for release candidates (with N being a natural number),
 % and 'stable' for stable versions
 
-'$lgt_version_data'(logtalk(3, 22, 0, b02)).
+'$lgt_version_data'(logtalk(3, 22, 0, b03)).
 
 
 
@@ -9999,16 +9999,22 @@ create_logtalk_flag(Flag, Value, Options) :-
 	Alias =.. [_| Args],
 	% allow for runtime use by adding a local definition that calls the remote
 	% definition except when the remote is a built-in predicate in "user" with
-	% no alias being defined
+	% no alias being defined or a built-in method that would clash with the
+	% local definition
 	'$lgt_comp_ctx'(Ctx, _, _, _, _, _, _, _, _, _, _, Mode, _, Lines),
 	(	Obj == user,
 		OriginalFunctor == AliasFunctor,
 		'$lgt_predicate_property'(Original, built_in) ->
+		% no need for a local definition
 		true
-	;	'$lgt_compile_aux_clauses'([(Alias :- Obj::Original)]),
-		% ensure that the uses/2 directive is found when looking for senders of this message
-		'$lgt_add_referenced_object_message'(Mode, Obj, Original, Alias, Alias)
+	;	'$lgt_built_in_method'(Alias, _, _, _) ->
+		% local definition would clash with a built-in method
+		true
+	;	% safe to add local definition
+		'$lgt_compile_aux_clauses'([(Alias :- Obj::Original)])
 	),
+	% ensure that the uses/2 directive is found when looking for senders of this message
+	'$lgt_add_referenced_object_message'(Mode, Obj, Original, Alias, Alias),
 	assertz('$lgt_pp_uses_predicate_'(Obj, Original, Alias, Lines)).
 
 '$lgt_compile_uses_directive_predicate_resource'(_, AliasFunctor, Arity, _, _) :-
@@ -10114,8 +10120,21 @@ create_logtalk_flag(Flag, Value, Options) :-
 	% unify arguments of TOriginal and TAlias
 	Original =.. [_| Args],
 	Alias =.. [_| Args],
-	% allow for runtime use by adding a local definition that calls the remote definition
-	'$lgt_compile_aux_clauses'([(Alias :- ':'(Module, Original))]),
+	% allow for runtime use by adding a local definition that calls the remote
+	% definition except when the remote is a built-in predicate in "user" with
+	% no alias being defined or a built-in method that would clash with the
+	% local definition
+	(	Module == user,
+		OriginalFunctor == AliasFunctor,
+		'$lgt_predicate_property'(Original, built_in) ->
+		% no need for a local definition
+		true
+	;	'$lgt_built_in_method'(Alias, _, _, _) ->
+		% local definition would clash with a built-in method
+		true
+	;	% safe to add local definition
+		'$lgt_compile_aux_clauses'([(Alias :- ':'(Module, Original))])
+	),
 	% ensure that the this use_module/2 directive is found when looking for callers of this module predicate
 	'$lgt_comp_ctx'(Ctx, _, _, _, _, _, _, _, _, _, _, Mode, _, Lines),
 	'$lgt_add_referenced_module_predicate'(Mode, Module, Original, Alias, Alias),